pyproject.toml
[tool.poetry]
name = "english-words-learning-bot"
version = "0.1.0"
description = ""
authors = ["Mikhail Egorov <pe.egorov@gmail.com>"]
readme = "README.md"
packages = [{include = "english_words_learning_bot"}]


[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.8.0"
asyncpg = "^0.29.0"
asyncio = "^3.4.3"
wordfreq = "^3.1.1"
setuptools = "^70.1.0"
python-dotenv = "^1.0.1"
nltk = "^3.8.1"
pyenchant = "^3.2.2"
requests = "^2.32.3"
deepl = "^1.18.0"
spacy = "^3.7.5"
textblob = "^0.18.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
